Split SerializedExecutor into SerializedInvoker
Use the SerializedInvoker in HttpChannel to avoid infinite recursion, but directly run tasks to avoid extra dispatch.
